Cube_Dodge v0.1.3
Added the blue block and implemented collision detection with it to
bounce the cubes off it. Also implemented the bouncing off the cubes.
Next time I need to work on the cubes flying off in the distance (minor
error in the bouncing). I also need to separate the collidable mesh list
so that every cube doesn't bounce with one collision.